\name{comclim-package}
\alias{comclim-package}
\alias{comclim}
\docType{package}
\title{
Community climate analysis
}
\description{
Computes community climate statistics for volume and mismatch using species' climate niches either unscaled or scaled relative to a regional species pool. These statistics can be used to describe biogeographic patterns and infer community assembly processes.
}
<<<<<<< HEAD
=======
\details{

}
>>>>>>> 006d8430e282886f59e5807a093be07648b08783
\author{
Benjamin Blonder <bblonder@gmail.com>
}
\references{
Blonder, B., Nogues-Bravo, D., et al. Linking environmental filtering and disequilibrium to biogeography with a community climate framework \emph{Ecology} (2014).
}
\keyword{ package }
