\name{summary.CommunityClimateInput}
\alias{summary.CommunityClimateInput}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summarizes community climate input
}
\description{Summarizes community climate input }
\usage{
\method{summary}{CommunityClimateInput}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
The object to be summarized
}
  \item{\dots}{
Other arguments to be passed
}
}