\name{summary.CommunityClimateStatistics}
\alias{summary.CommunityClimateStatistics}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summarizes community climate statistics
}
\description{Summarizes community climate statistics}
\usage{
\method{summary}{CommunityClimateStatistics}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
The object to be summarized
}
  \item{\dots}{
Other arguments to be passed
}
}