\name{comclim-package}
\alias{comclim-package}
\alias{comclim}
\docType{package}
\title{
Community climate analysis
}
\description{
Computes community climate statistics for volume and mismatch using species' climate niches either unscaled or scaled relative to a regional species pool. These statistics can be used to describe biogeographic patterns and infer community assembly processes.
}
\details{
\tabular{ll}{
Package: \tab comclim\cr
Type: \tab Package\cr
Version: \tab 0.9.4\cr
Date: \tab 2014-18-9\cr
License: \tab GPL-3\cr
Depends: \tab methods\cr
}
}
\author{
Benjamin Blonder <bblonder@gmail.com>
}
\references{
Blonder, B., Nogues-Bravo, D., et al. Linking environmental filtering and disequilibrium to biogeography with a community climate framework \emph{Ecology} (2014).
}
\keyword{ package }
