\name{comclim-package}
\alias{comclim-package}
\alias{comclim}
\docType{package}
\title{
Community climate analysis
}
\description{
Computes community climate statistics for volume (Delta) and lag (Lambda) using species' climate niches. These statistics can be used to infer community assembly processes of environmental filtering and dispersal limitation / species interactions respectively. Includes a vignette outlining usage.
}
\details{
\tabular{ll}{
Package: \tab comclim\cr
Type: \tab Package\cr
Version: \tab 0.9.1\cr
Date: \tab 2013-11-25\cr
License: \tab GPL-3\cr
Depends: \tab methods\cr
}
}
\author{
Benjamin Blonder <bblonder@gmail.com>
}
\references{
Blonder, B., Nogues-Bravo, D., et al. Assembly of communities in climate space. \emph{submitted}.
}
\keyword{ package }
